Move the configuration into the infrasturcture project.
